+++ tests/src/test/java/testdatasource/account-data.json
@@ -5,6 +5,6 @@
     "nonExistingCSOGuid": "88da92db-0791-491e-8c58-1a969e67d2fb",
     "validUserId": "90a5e1d5-9578-4573-b965-02f870780b1f",
     "inValidUserId": "A02F2D0F367E4575BC782A414DC9369D",
-    "clientSecret": ""
+    "clientSecret": "<clientSecret>"
   }
 ]
